… libobs/obs-video.c