@jpark37 2019-07-27
I422 is not supported by OBS as an output format.