@energizerfellow 2019-07-27
True, but should it be added at some point with 4:2:2's prevalence in the pro video world?