@jpark37 2019-07-27
Finished testing, and I verified both Mantis bugs are fixed.