@Xaymar 2019-07-27
I'm pretty sure there are 4:2:2 options for use in encoders:

```
	/* packed 422 formats */
	VIDEO_FORMAT_YVYU,
	VIDEO_FORMAT_YUY2, /* YUYV */
	VIDEO_FORMAT_UYVY,
```

I'm using these in obs-ffmpeg-encoder and enc-vfw.