@jpark37 2019-07-27
OBS has shaders to convert 4:2:2 YCbCr to RGB, but not the reverse. There are only 4 options in the output settings: NV12, I420, I444, and RGB.