@jpark37 2019-07-28
This would also make #1655 unnecessary because those shader functions have been deleted.