@SuslikV 2019-09-21
> ... I420/NV12 formats now have chroma shifted to
> the left as 4:2:0 is shown in the H.264 specification

Document needs clarification with year of the last redaction used as base source for the implementation.